Pioneer bandit class and I think a bomb vendor
I have no idea what the fuck the bomb vendor is but it's apparently necessary for the pr I think, someone pls help. Also rebalances the pulaski axe to be somewhat on par but not really better than the warden axe (FUCK THEIR SNOWFLAKE ROLE!!!) but yeah adds a new bandit class called pioneer but a better name would be trapper or sapper or some shit because it has traps and bombs, basically your steppe sapper if he lost his axe and decided shovels were cooler.